Switching comparison statistics for the Metrics to a new MetricsUtilities.
